\name{predict.cslogit}

\alias{predict.cslogit}

\title{Predict method for cslogit fits}

\description{Obtains predictions from a fitted \code{cslogit} object.}

\usage{
\method{predict}{cslogit}(object, newdata = NULL, \dots)
}

\arguments{
\item{object}{a fitted object of class "\code{cslogit}".}
\item{newdata}{optionally, a data frame in which to look for variables with which to predict. If omitted, the fitted coefficients are used.}
\item{...}{further arguments passed to or from other methods.}
}

\details{...}

\value{
\code{predict.cslogit} returns a vector of predicted probabilities.
}

\references{...}

\author{Sebastiaan Höppner}


\seealso{\code{\link{cslogit}}, \code{\link{predict}}}

\examples{
# Provide an example (TODO).
}
