\name{summary.cslogit}

\alias{summary.cslogit}

\title{Summarizing cslogit fits}

\description{The function is a method for class \code{cslogit} objects that summarizes their fit.}

\usage{
\method{summary}{cslogit}(object, \dots)
}


\arguments{
\item{object}{an object of class "\code{cslogit}". A result of a call to the \code{cslogit} function.}
\item{...}{further arguments passed to or from other methods.}
}

\details{\code{summary.cslogit} prints a summary of the "\code{cslogit}" object.}

\references{...}

\author{Sebastiaan Höppner}


\seealso{\code{\link{cslogit}}, \code{\link{summary}}}

\examples{
# Provide an example (TODO).
}
